requirements.txt
capsolver_python
capmonster_python
python-dotenv
tls_client
PyYAML